% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomdm.R
\name{binomdm}
\alias{binomdm}
\title{DataMaker for data following the binomial distribution}
\usage{
binomdm(k, n, groupnames, size, prob)
}
\arguments{
\item{k}{number of groups that you have tested}

\item{n}{sample size for each of the groups}

\item{groupnames}{vector of names for the groups that you tested; make sure they are in order}

\item{size}{the number of successes per group}

\item{prob}{probability of success for each group}

\item{plot}{TRUE or FALSE, to generate jitterplot}
}
\value{
a list of the input parameters and simulated data
}
\description{
DataMaker for data following the binomial distribution
}
\examples{
binomdm(k = 3 , groupnames = c("control", "treatment1", "treatment2")
, n = c(40, 55, 55), size = c(5, 20, 20), prob = c(0.2, 0.2, 0.2))


}
