% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormdm.R
\name{lnormdm}
\alias{lnormdm}
\title{Lognormal Data Maker : lnormdm}
\usage{
lnormdm(k, names = c(...), n = c(...), meanlog = c(...), sdlog = c(...))
}
\arguments{
\item{k}{is the number of treatment groups.}

\item{names}{are the names of the treatment groups.}

\item{n}{is the size of each treatment group.}

\item{meanlog}{is the log mean of each treatment group.}

\item{sdlog}{is the log standard deviation of each treatment group.}
}
\value{
ggplot,data
}
\description{
A sandbox to create and plot lognormal data.
Enter any number of treatment groups, their names, their respective log means, and their log standard deviations.
Treatment groups can be uneven sizes.
k must correspond with the number of values in names, meanlog, and sdlog.
}
