% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinomdm.R
\name{nbinomdm}
\alias{nbinomdm}
\title{Negative Binomial Data Maker}
\usage{
nbinomdm(k, names, n, size, mu, hist, scatter)
}
\arguments{
\item{k}{A single value corresponding to the number of treatment groups.}

\item{names}{A vector of character strings specifying the names of the treatment groups.}

\item{n}{A vector specifying the sample sizes of the treatment groups.}

\item{size}{A vector specifying the dispersion parameters of the negative binomial distributions for the treatment groups. Strictly have to be positive values, not necessarily integers.}

\item{mu}{A vector specifying the means of the negative binomial distributions for the treatment groups.}

\item{hist}{TRUE or FALSE to return a histogram plot}

\item{scatter}{TRUE or FALSE to return a scatter plot}
}
\value{
Histogram and jitter plots for the treatment groups and a data frame containing the random samples.
}
\description{
nbinomdm simulates over-dispersed frequency data using the stats::rnbinom function.
It creates a data frame containing different samples for different treatment groups 
following the negative binomial distribution. Currently it uses the size and mu parameters
following the rnbinom function, but not the prob parameter.
}
\examples{
nbinomdm(3,c("C","T1","T2"),c(50,50,60),c(3,3,3),c(1,5,10),TRUE,FALSE)


}
