% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normdm.R
\name{normdm}
\alias{normdm}
\title{normdm}
\usage{
normdm(k, names = c(...), n = c(...), mean = c(...), sd = c(...))
}
\arguments{
\item{k}{number of groups to be simulated}

\item{names}{names for each group}

\item{n}{sample size of each group}

\item{mean}{mean of simulated data for each group}

\item{sd}{standard deviation of simulated data for each group}
}
\value{

}
\description{
normdm
}
\examples{
normdm(k=5, names= c("control", "low", "medium", "high", "positive"),n= c(50,50,50,50,40), mean = c(10, 5, 6, 7,8), sd= c(1,5,5,5,5))


}
