% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox.R
\name{wl.pwr}
\alias{wl.pwr}
\title{Wilcoxon Rank Sum & Signed Rank Tests Repeater Function}
\usage{
wl.pwr(
  sims,
  param,
  alternative,
  mu,
  paired,
  exact,
  correct,
  conf.int,
  conf.level,
  tol.root,
  digits.rank
)
}
\arguments{
\item{sims}{number of simulations}

\item{param}{parameters from datamaker output}

\item{alternative}{"greater", "less", or "two.sided"}

\item{mu}{null hypothesis}

\item{paired}{paired test, T/F}

\item{exact}{should exact pvalue be computed, T/F}

\item{correct}{should continuity correction be applied to pvalue, T/F}

\item{conf.int}{should confidence interval be computed, T/F}

\item{conf.level}{confidence level of the interval}

\item{tol.root}{if conf.int=T, positive numeric tolerance}

\item{digits.rank}{used to computed ranks of test statistic}
}
\value{
a power value
}
\description{
Wilcoxon Rank Sum & Signed Rank Tests Repeater Function
}
\examples{
wl.pwr(sims=10000,param=d1$param,alternative="two.sided",mu=0,paired=FALSE,exact=NULL,correct=TRUE,conf.int=FALSE,conf.level=0.95,tol.root=1e-4,digits.rank=Inf)
}
