% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinomdm.R
\name{nbinomdm}
\alias{nbinomdm}
\title{Negative Binomial Data Maker}
\usage{
nbinomdm(k, names, n, size, mu)
}
\arguments{
\item{k}{A single value corresponding to the number of treatment groups.}

\item{names}{A vector of character strings specifying the names of the treatment groups.}

\item{n}{A vector specifying the sample sizes of the treatment groups.}

\item{size}{A vector specifying the dispersion parameters of the negative binomial distributions for the treatment groups. Strictly have to be positive values, not necessarily integers.}

\item{mu}{A vector specifying the means of the negative binomial distributions for the treatment groups.}
}
\value{
Histogram and jitter plots for the treatment groups and a data frame containing the random samples.
}
\description{
nbinomdm uses rnbinom function from the stats package to create a data frame
containing different samples for different treatment groups following the
negative binomial distribution. Currently it uses the size and mu parameters
following the rnbinom function, but not the prob parameter.
}
\examples{
nbinomdm(k=3,names=c("Control","Treatment1","Treatment2"),n=c(50,50,60),size=c(3,3,3),mu=c(5,10,20))


}
