% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisdm.R
\name{poisdm}
\alias{poisdm}
\title{Poisson DataMaker: simulates groups of Poisson-distributed count data.}
\usage{
poisdm(n, lambda, groupnames, plot)
}
\arguments{
\item{n}{sample size}

\item{lambda}{group mean}

\item{groupnames}{names of the groups}

\item{plot}{TRUE or FALSE; generate histogram and jitterplot of output}
}
\value{
a data frame with input values; if plot == T, also a histogram, and a jitterplot
}
\description{
Poisson DataMaker: simulates groups of Poisson-distributed count data.
}
\examples{
poisdm(c(40, 50), c(10, 12), c("group1", "group2"), TRUE)
}
