% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisdm.R
\name{poisdm}
\alias{poisdm}
\title{Poisson datamaker}
\usage{
poisdm(n, lambda, groupnames)
}
\arguments{
\item{n}{sample size}

\item{lambda}{group mean}

\item{groupnames}{names of the groups}
}
\value{

}
\description{
Poisson datamaker
}
\examples{
poisdm(c(40, 50), c(10, 12), c("group1", "group2"))
}
